refactor: move augment and label into data package
