+++ dfode_kit/cli/commands/augment.py
@@ -48,8 +48,7 @@ def add_command_parser(subparsers):
 def handle_command(args):
     import numpy as np
 
-    from dfode_kit.data_operations.augment_data import random_perturb
-    from dfode_kit.data.io_hdf5 import get_TPY_from_h5
+    from dfode_kit.data import get_TPY_from_h5, random_perturb
 
     print('Handling augment command')
     print(f'Loading data from h5 file: {args.h5_file}')
