--- dfode_kit/data_operations/contracts.py
@@ -1,1 +1,0 @@
-from dfode_kit.data.contracts import *  # noqa: F401,F403
