--- dfode_kit/data_operations/h5_kit.py
@@ -1,37 +1,0 @@
-"""Compatibility shim for the canonical data I/O and integration modules."""
-
-from importlib import import_module
-
-
-__all__ = [
-    "touch_h5",
-    "get_TPY_from_h5",
-    "advance_reactor",
-    "load_model",
-    "predict_Y",
-    "nn_integrate",
-    "integrate_h5",
-    "calculate_error",
-]
-
-_ATTRIBUTE_MODULES = {
-    "touch_h5": ("dfode_kit.data.io_hdf5", "touch_h5"),
-    "get_TPY_from_h5": ("dfode_kit.data.io_hdf5", "get_TPY_from_h5"),
-    "advance_reactor": ("dfode_kit.data.integration", "advance_reactor"),
-    "load_model": ("dfode_kit.data.integration", "load_model"),
-    "predict_Y": ("dfode_kit.data.integration", "predict_Y"),
-    "nn_integrate": ("dfode_kit.data.integration", "nn_integrate"),
-    "integrate_h5": ("dfode_kit.data.integration", "integrate_h5"),
-    "calculate_error": ("dfode_kit.data.integration", "calculate_error"),
-}
-
-
-def __getattr__(name):
-    if name not in _ATTRIBUTE_MODULES:
-        raise AttributeError(f"module 'dfode_kit.data_operations.h5_kit' has no attribute '{name}'")
-
-    module_name, attribute_name = _ATTRIBUTE_MODULES[name]
-    module = import_module(module_name)
-    value = getattr(module, attribute_name)
-    globals()[name] = value
-    return value
