+++ tests/test_data_integration_shims.py
@@ -1,17 +1,12 @@
 import importlib
 
 
-def test_h5_kit_shim_reexports_io_helpers():
-    legacy_h5_kit = importlib.import_module("dfode_kit.data_operations.h5_kit")
+def test_data_integration_module_exports_expected_helpers():
+    integration = importlib.import_module("dfode_kit.data.integration")
     canonical_io = importlib.import_module("dfode_kit.data.io_hdf5")
 
-    assert legacy_h5_kit.touch_h5 is canonical_io.touch_h5
-    assert legacy_h5_kit.get_TPY_from_h5 is canonical_io.get_TPY_from_h5
-
-
-def test_data_operations_package_reexports_canonical_io_helpers():
-    legacy_data_ops = importlib.import_module("dfode_kit.data_operations")
-    canonical_io = importlib.import_module("dfode_kit.data.io_hdf5")
-
-    assert legacy_data_ops.touch_h5 is canonical_io.touch_h5
-    assert legacy_data_ops.get_TPY_from_h5 is canonical_io.get_TPY_from_h5
+    assert callable(integration.advance_reactor)
+    assert callable(integration.integrate_h5)
+    assert callable(integration.calculate_error)
+    assert callable(canonical_io.touch_h5)
+    assert callable(canonical_io.get_TPY_from_h5)
